+++ control_server/changelog.md
@@ -7,6 +7,7 @@ __*Constellation*__ 4 expands customization for your apps and provides many qual
 
 ### General
 - Constellation Apps now checks for software updates in standalone mode.
+- In addition to a solid color, you can now set app backgrounds to an image or a two-color gradient.
 
 # **Constellation 3.3**
 __*Constellation*__ 3.3 expands configuration options for many apps and further streamlines the user experience on Windows.